data/automatick.py
import pandas as pd
import json

def process_location_data(input_file, output_file):
    """
    Process Indian states, districts, talukas, and villages into a JSON structure.
    :param input_file: Path to the Excel or CSV file containing the location data.
    :param output_file: Path to the output JSON file.
    """
    try:
        # Load the data
        data = pd.read_excel(input_file)  # Use read_csv if the file is in CSV format
    except FileNotFoundError:
        print(f"File not found: {input_file}")
        print("Please ensure the file exists and provide the correct path.")
        return
    except Exception as e:
        print(f"Error reading file: {e}")
        return

    # Ensure required columns exist
    required_columns = ["State Name", "District Name", "Taluka Name", "Village Name"]
    if not all(col in data.columns for col in required_columns):
        print(f"Input file must contain the following columns: {required_columns}")
        return

    # Process the data...

