requirements.txt
PyQt5==5.15.9
PyQtChart==5.15.6
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pyinstaller==6.3.0
PyGithub==2.1.1
customtkinter==5.2.1
pillow==10.1.0
reportlab==10.0.0
cairosvg==2.7.1
